+++ resources/views/docs/partials/modal-livewire-preview.blade.php
@@ -1,0 +1,1 @@
+@livewire(\App\Livewire\Docs\ModalPreview::class, ['title' => $title], key('modal-preview-'.$index))